refactor: add calculate form component
